# src/swagger.json
{
  "basePath": "/",
  "consumes": ["application/json"],
  "info": {
    "title": "Passphrase Launchpad Server",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    }
  },
...
  "paths": {
...
    "/v1/fundingPools/search": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "searchFundingPools",
        "summary": "Search FundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/FundingPools"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchFundignPools",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          }
        ]
      }
    },
    "/v1/fundingPools/{id}/addFundraisingSC": {
      "post": {
        "tags": ["FundingPools"],
        "operationId": "addFundraisingSC",
        "summary": "Add Fundraising SC",
        "description": "Endpoint for add fundraising SC",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful added",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "addFundraisingSC",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "contractAddress": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v2/fundingPools/search": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "searchFundingPools",
        "summary": "Search FundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/FundingPools"
                  }
                },
                "count": {
                  "type": "number",
                  "default": 0
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "searchFundignPools",
              "type": "object",
              "required": ["success"]
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": true,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 0
          }
        ]
      }
    },
    "/v2/fundingPools/{id}/addFundraisingSC": {
      "post": {
        "tags": ["FundingPools"],
        "operationId": "addFundraisingSC",
        "summary": "Add Fundraising SC",
        "description": "Endpoint for add fundraising SC",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful added",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "addFundraisingSC",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "properties": {
                "id": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["id"]
            }
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "contractAddress": {
                  "type": "string",
                  "default": null
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/fundingPools": {
      "post": {
        "tags": ["FundingPools"],
        "operationId": "createFundingPool",
        "summary": "Create Funding Pool",
        "description": "Endpoint for create funding pool",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful created",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "createFundingPool",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getAllFundingPools",
        "summary": "Get All Funding Pools",
        "description": "Endpoint for get all fundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object",
              "required": ["id"]
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": []
      }
    },
    "/v2/fundingPools": {
      "post": {
        "tags": ["FundingPools"],
        "operationId": "createFundingPool",
        "summary": "Create Funding Pool",
        "description": "Endpoint for create funding pool",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful created",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "createFundingPool",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getAllFundingPools",
        "summary": "Get All Funding Pools",
        "description": "Endpoint for get all fundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object",
              "required": ["id"]
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": []
      }
    },
    "/v1/fundingPools/projects": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getFilteredPools",
        "summary": "Get Filtered Pools",
        "description": "Endpoint for get filtered pools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object",
              "required": ["id"]
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": []
      }
    },
    "/v2/fundingPools/projects": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getFilteredPools",
        "summary": "Get Filtered Pools",
        "description": "Endpoint for get filtered pools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object",
              "required": ["id"]
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": []
      }
    },
    "/v1/fundingPools/{slug}": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getFundingPoolBySlug",
        "summary": "Get Funding Pool By Slug",
        "description": "Endpoint for get filtered pools by slug",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object",
              "required": ["id"]
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "schema": {
              "properties": {
                "slug": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["slug"]
            }
          }
        ]
      }
    },
    "/v2/fundingPools/{slug}": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getFundingPoolBySlug",
        "summary": "Get Funding Pool By Slug",
        "description": "Endpoint for get filtered pools by slug",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object",
              "required": ["id"]
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "schema": {
              "properties": {
                "slug": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["slug"]
            }
          }
        ]
      }
    },
    "/v1/fundingPools/company/{companyId}": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getCompanyFundingPools",
        "summary": "Get Company FundingPools",
        "description": "Endpoint for get company fundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "schema": {
              "properties": {
                "companyId": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["companyId"]
            }
          }
        ]
      }
    },
    "/v2/fundingPools/company/{companyId}": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getCompanyFundingPools",
        "summary": "Get Company FundingPools",
        "description": "Endpoint for get company fundingPools",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "$ref": "#/definitions/FundingPools",
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "schema": {
              "properties": {
                "companyId": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["companyId"]
            }
          }
        ]
      }
    },
    "/v1/fundingPools/{fundingPoolId}/getAllSafts": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getFoundingPoolSafts",
        "summary": "Get Founding Pool Safts",
        "description": "Endpoint for get founding pool safts",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "properties": {
                "saftId": {
                  "type": "string",
                  "default": null
                },
                "procedureId": {
                  "type": "string",
                  "default": null
                },
                "signatureId": {
                  "type": "string",
                  "default": null
                },
                "ownerId": {
                  "type": "string",
                  "default": null
                },
                "isValid": {
                  "type": "boolean",
                  "default": false
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "fundingPoolId",
            "schema": {
              "properties": {
                "fundingPoolId": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["fundingPoolId"]
            }
          }
        ]
      }
    },
    "/v2/fundingPools/{fundingPoolId}/getAllSafts": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getFoundingPoolSafts",
        "summary": "Get Founding Pool Safts",
        "description": "Endpoint for get founding pool safts",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "properties": {
                "saftId": {
                  "type": "string",
                  "default": null
                },
                "procedureId": {
                  "type": "string",
                  "default": null
                },
                "signatureId": {
                  "type": "string",
                  "default": null
                },
                "ownerId": {
                  "type": "string",
                  "default": null
                },
                "isValid": {
                  "type": "boolean",
                  "default": false
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "fundingPoolId",
            "schema": {
              "properties": {
                "fundingPoolId": {
                  "type": "string"
                }
              },
              "title": "CreatePath",
              "type": "object",
              "required": ["fundingPoolId"]
            }
          }
        ]
      }
    },
    "/v1/fundingPools/pastDeals/companies": {
      "get": {
        "tags": ["FundingPools"],
        "operationId": "getPastDeals",
        "summary": "Get Past Deals",
        "description": "Endpoint for get past deals",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successfully got",
            "schema": {
              "properties": {
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "default": null
                      },
                      "companies": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "iconUrl": {
                              "type": "string",
                              "default": null
                            },
                            "name": {
                              "type": "string",
                              "default": null
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "success": {
                  "type": "string",
                  "default": true
                }
              },
              "title": "CreateBody",
              "type": "object"
            }
          },
          "401": {
            "description": "Something went wrong"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": []
      }
    },
...
  },
  "produces": ["application/json"],
  "swagger": "2.0",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "body",
      "description": "This API requires authentication using a bearer token. Please include the token in the 'Authorization' header of your requests. ex: Bearer [space] <your-token>"
    }
  }
}